+++ dio/CHANGELOG.md
@@ -6,6 +6,8 @@ See the [Migration Guide][] for the complete breaking changes list.**
 ## Unreleased
 
 - Fix failing requests throw `DioException`s with `.unknown` instead of `.connectionError` on `SocketException`.
+- Fix wrong formatting of multi-value header in `BrowserHttpClientAdapter`.
+- Add warning in debug mode when trying to send data with a `GET` request in web.
 
 ## 5.3.2